+++ NEWS.md
@@ -1,3 +1,5 @@
+# C50 (development version)
+
 # C50 0.2.0
 
 * Updated C code to stop overwriting keyword symbols.